Move the jmx prometheus jar into the correct build location.
Continuing to iron out problems with the migration to using RHEL8 images. After seeing problems with the python binary missing in PATH, we're now in familiar territory where the jmx_prometheus_javaagent:0.3.1 jar file is missing when we attempt to load that, which is something we also saw in the Hadoop image. IT looks like there may be something wrong with our maven environment, but my understanding is that maven isn't respecting the `-Ddest` flag value, and instead is storing that downloaded jar into the default ~/.m2/repository location. If we move that jar from that directory to the expected path that the Dockerfile is configured to use, this error should discontinue.