Dockerfile: Avoid using the `alternatives` command when building outside of CI
Currently the Dockerfile.rhel is configured to use RHEL8-based images whereas ART hasn't made the migration over from RHEL7 to RHEL8 yet. In order to abstract the linkage of the python3 rpm to the default python path, we use the `alternatives` command, which seems to be non-existent in the default
RHEL7 image that ART is configured to use. As a result, CI is able to successfully build presto RHEL images, whereas ART is failing to build as `alteratives` is an undefined command.