Dockerfile: Remove conditional for symlinking python3 to python
This conditional was needed as there was still a gap between our CI images, which are RHEL8-based, and the ART images, which was still on RHEL7.

Now that it appears the switch has been made, we can remove the conditional and always use the `alternatives` command to symlink the python3 rpm to the python command.